Finished button and fixed stuff
Implemented a cooldown timer for the white blood cell button.

Implemented the deploy white blood cell ability.

Fixed a bug that caused a white blood cell to be fired if the mouse was clicked outside the bounds of the button.

Fixed a bug that prevented players from firing a white blood cell after another white blood cell had been destroyed.

Added a unique tag to the deployed white blood cell prefab and included it in the conditions for the OnTriggerEnter2D function within the Projectile_Destruction script.

Disabled the old button script.

Increased the size of the doctor by a factor of 3.