Made some changes to transition
transition now is on its own canvas and is on a higher sorting layer than our normal UI. UI, with the exception of the joystick due to some weird glitches, are disabled during the rest puzzle menu (some manual code needed their--to be polished later). LateUpdate added to overide a problem where the ui woulnt be disabled in overworld. !!!!Transitions still need to be added on other scenes !!!.